pyproject.toml
[project]
name = "partsuite"
version = "0.1.0"
description = "Invoice parsing demo with FastAPI and Vite UI"
authors = [{name = "Auto Developer"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn>=0.29",
  "sqlalchemy>=2.0",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pdfminer.six>=20231228",
  "pypdf>=4.0",
  "python-multipart>=0.0.9",
  "pytest>=8.2",
  "httpx>=0.27"
]

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
